fix(nuxt): Re-export 'default' exports with rollup plugin (#13984)
While the preset for Netlify exports the serverless handler function as
`export { D as default }`, the Vercel preset exports this handler as
`export { D as handler };`.

This PR makes some adaptions to the code generation in the plugin to
make it possible to re-export `default` functions. The previous snippet
did not work as `default` is obviously not allowed as a function name.